+++ app-Prefs/ViewController.swift
@@ -252,14 +252,24 @@ extension ViewController: UITableViewDelegate, UITableViewDataSource {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         if !editClicked {
             var action: String?
+            var name: String?
             let realm = try! Realm()
             let model = realm.objects(Setting.self).filter("isDeleted = false && sortNum = \(indexPath.row)").first!
-            if model.type == ActionType.system.rawValue {
-                action = "app-Prefs:\(model.action!)"
-            } else {
+//            if model.type == ActionType.system.rawValue {
+//                action = "app-Prefs:\(model.action!)"
+//            } else {
                 action = model.action
-            }
+//            }
+            name = model.name
             
+            let displayStr = name! + ": " + action!
+            let alertC = UIAlertController(title: NSLocalizedString("Detail", comment: ""), message: displayStr, preferredStyle: .alert)
+            alertC.addAction(UIAlertAction(title: NSLocalizedString("Cancel", comment: ""), style: .cancel, handler: nil))
+            
+            present(alertC, animated: true, completion: nil)
+            
+            
+            /**
             UIApplication.shared.open(URL.init(string: action!)!, options: [:]) { (ret) in
                 if ret == false {
 
@@ -275,6 +285,7 @@ extension ViewController: UITableViewDelegate, UITableViewDataSource {
                 }
                 
             }
+            */
 
         } else {